Refactor solution detail codec and SolutionTemplateProps
