+++ apps/nextjs-website/src/app/solutions/[solutionSlug]/page.tsx
@@ -3,14 +3,14 @@ import { Metadata } from 'next';
 import { baseUrl } from '@/config';
 import { getSolution } from '@/lib/api';
 import SolutionTemplate from '@/components/templates/SolutionTemplate/SolutionTemplate';
-import { getFullSolutionsProps } from '@/lib/cmsApi';
+import { getSolutionsProps } from '@/lib/cmsApi';
 
 type Params = {
   solutionSlug: string;
 };
 
 export async function generateStaticParams() {
-  const solutions = await getFullSolutionsProps();
+  const solutions = await getSolutionsProps();
   return [...solutions].map(({ slug }) => ({
     solutionSlug: slug,
   }));
