+++ apps/nextjs-website/src/components/templates/SolutionTemplate/SolutionTemplate.tsx
@@ -51,6 +51,8 @@ export type SolutionTemplateProps = {
       };
     }[];
   };
+  solutionSlug: string;
+  path?: string;
 };
 
 const SolutionTemplate = ({
