--- apps/nextjs-website/src/lib/types/apiData.ts
@@ -1,4 +1,0 @@
-import { ApiPageProps } from '@/app/[productSlug]/api/[apiDataSlug]/page';
-import { Page } from '@/lib/types/page';
-
-export type ApiData = Page<Omit<ApiPageProps, 'products'>>;
