--- apps/nextjs-website/src/lib/types/solution.ts
@@ -1,11 +1,0 @@
-export type Solution = {
-  readonly slug: string;
-  readonly kickerTitle: string;
-  readonly title: string;
-  readonly description?: string;
-  readonly dirName: string;
-  readonly landingUseCaseFile: string;
-  readonly publishedAt: Date;
-  readonly icon: string;
-  readonly introductionToSteps?: string;
-};
